Update Ty Gwyn Game (Phase 3 - Exit the room) - Testing Nesting.py
getting there now. the room logic is in place and the battle logic has been added. Now we need to add dialogue for random mobs, drinking / drunk logic and we should be ready to go! 

all the logic I have added since the last commit is untested so will need testing and amending first